Add support to encode/decode struct pointers
The zero-value of a section is not always enough to determine whether
or not a section should be present or not. For example, you could want
to explicitly set a property to false because the default is true when
absent. In this case, it becomes the zero value of the section which
prevents the section from being written.